pyproject.toml
[tool.poetry]
name = "marazzo-ecommerce-website"
version = "0.1.0"
description = ""
authors = ["Takib Yeasar <mdtakibyeasar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.3.0"
psycopg2 = "^2.9.9"
setuptools = "^68.2.2"
pillow = "^10.1.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
